requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
sqlparse==0.4.4
typing_extensions==4.11.0
